+++ src/main/java/umc/spring/service/review/ReviewQueryService.java
@@ -1,0 +1,9 @@
+package umc.spring.service.review;
+
+import org.springframework.data.domain.Page;
+import umc.spring.domain.Review;
+
+public interface ReviewQueryService {
+
+    Page<Review> getMyReviewList(Integer page);
+}
