+++ src/main/java/umc/spring/web/controller/ReviewController.java
@@ -15,6 +15,7 @@ import umc.spring.converter.ReviewConverter;
 import umc.spring.domain.Review;
 import umc.spring.service.member.MemberQueryService;
 import umc.spring.service.review.ReviewCommandService;
+import umc.spring.service.review.ReviewQueryService;
 import umc.spring.validation.annotation.CheckPage;
 import umc.spring.web.dto.MissionResponseDto;
 import umc.spring.web.dto.ReviewRequestDto;
@@ -26,7 +27,7 @@ import umc.spring.web.dto.StoreResponseDto;
 public class ReviewController {
 
     private final ReviewCommandService reviewCommandService;
-    private final MemberQueryService memberQueryService;
+    private final ReviewQueryService reviewQueryService;
 
     @PostMapping("/reviews")
     public ApiResponse<ReviewResponseDto.addResultDto> addReview (@RequestBody @Valid ReviewRequestDto.addReviewDto request) {
@@ -46,7 +47,7 @@ public class ReviewController {
     })
     public ApiResponse<MissionResponseDto.reviewReviewListDto> getReviewList (@CheckPage @RequestParam(name = "page") Integer page){
 
-        Page<Review> getMyReviewList = memberQueryService.getMyReviewList(page-1);
+        Page<Review> getMyReviewList = reviewQueryService.getMyReviewList(page-1);
 
         return ApiResponse.onSuccess(MissionConverter.reviewReviewListDto(getMyReviewList));
     }
